ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
pillow==10.1.0
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.26.4
watchfiles==0.21.0